requirements.txt
boto3
paramiko
psycopg2-binary
tox